--- packages/fkweb/create/urls.py
@@ -1,8 +1,0 @@
-from django.conf.urls import include
-from django.conf.urls import url
-
-from create import views
-
-urlpatterns = [
-    url(r'^$', views.home, name='create_home'),
-]
