+++ packages/fkweb/fkbeta/settings/base.py
@@ -245,8 +245,6 @@ THIRD_PARTY_APPS = (
 
 # Apps specific for this project go here.
 LOCAL_APPS = (
-    #'fktasks',
-    'create',
     'fkbeta',
     'agenda',
     'fk',
