--- packages/fkweb/fkvod/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (c) 2012-2013 Benjamin Bruheim <grolgh@gmail.com>
-# This file is covered by the LGPLv3 or later, read COPYING for details.
