--- packages/fkweb/fkvod/tests.py
@@ -1,24 +1,0 @@
-# Copyright (c) 2018 Petter Reinholdtsen <pere@hungry.com>
-# This file is covered by the LGPLv3 or later, read COPYING for details.
-
-from django.test import TestCase
-from django.urls import reverse
-
-class VodTests(TestCase):
-    fixtures = ['test.yaml']
-
-    def test_video_list(self):
-        response = self.client.get(reverse('vod-video-list'))
-        self.assertContains(response, 'RSS', count=1)
-
-    def test_video_detail(self):
-        response = self.client.get(reverse('vod-video-detail', args=[1]))
-        self.assertContains(response, 'Category', count=1)
-
-    def test_category_list(self):
-        response = self.client.get(reverse('vod-category-list'))
-        self.assertContains(response, 'All Categories')
-
-    def test_org_list(self):
-        response = self.client.get(reverse('vod-org-video-list', args=[1]))
-        self.assertContains(response, 'Videos from')
