--- packages/fkweb/templates/fkvod/video_details.html
@@ -1,95 +1,0 @@
-{% extends "base_generic.html" %}
-{% load vod %}
-{% load fk_utils %}
-{% load static %}
-
-{% block style %}
-  <style>
-  .video_container {
-    overflow:auto;
-  }
-
-  .video-info {
-    overflow:auto;
-    margin: 0px 16px 0px 16px;
-    padding: 16px;
-    border-radius: 8px;
-    background-color:#eaeaea;
-    min-width: 100px;
-  }
-
-  .video-description {
-    overflow:auto;
-    margin: 0px 16px 0px 16px;
-    padding: 8px 16px 8px 16px;
-    border-radius: 8px;
-    background-color:#eaeaea;
-    font-size:120%;
-    clear:both;
-  }
-  </style>
-{% endblock style %}
-
-{% block content %}
-  <div class="video-container">
-    <div class="widget-container">
-      {% show_vod_widget video_id %}
-    </div>
-    <div class="video-info">
-      {% if video.organization %}
-        <div>
-          <a href="{% url 'vod-org-video-list' video.organization.id %}">{{ video.organization }}</a>
-        </div>
-      {% endif %}
-      {% if video.ref_url %}
-        <div>
-          <a href="{{ video.ref_url }}">{{ video.ref_url }}</a>
-       </div>
-      {% endif %}
-      <div>
-        {% trans "Uploaded by:" %} <i>{{ video.creator.first_name }} {{ video.creator.last_name }}</i>
-      </div>
-      <div>
-        {% trans "Category:" %} <i>
-        {% for category in video.category_list %}
-          <span style="background-color:#ccc; border-radius: 4px; padding:2px;"><a href="{% url 'vod-category-video-list' category %}">{{ category }}</a></span>
-        {% endfor %}</i>
-      </div>
-      {% if video %}
-        <div>
-          <small>{% trans "Uploaded:" %} <i>{{ video.uploaded_time|date:"d M Y"}}</i></small>
-        </div>
-        <div>
-          <small>{% trans "Length:" %} {{ video.duration|truncate_duration }} {% if video.is_filler %}(filler){% endif %}
-        </div>
-      {% endif %}
-      <div>
-        {% if video.first_broadcast %}
-          <small>{% trans "First broadcast:" %} <i>{{ video.first_broadcast.starttime }}</i></small>
-        {% else %}
-          <small>{% trans "Not broadcast yet." %}</small>
-        {% endif %}
-        <!-- <div>Previous broadcast: <i> {{ video.last_broadcast.starttime }}</i></div> -->
-      </div>
-      {% if video.has_tono_records %}
-        <div><img src="{% static 'tono-logo.gif' %}" alt="TONO" style="margin:5px;"/></div>
-      {% endif %}
-      {% if allowed_to_edit %}
-	<a href="{% url 'manage-video-edit' video.id %}">{% trans "Edit metadata" %}</a>
-      {% endif %}
-    </div> {# .video-info #}
-  </div> {# .video-container #}
-  {% if video.header %}
-    <div class="video-description">
-      {{ video.header|linebreaks }}
-      <!--
-      <ul style="clear:both;">
-        <li><b>History</b></li>
-        {% for event in video.schedule  %}
-          <li>{{ event }}</li>
-        {% endfor %}
-      </ul>
-      -->
-    </div>
-  {% endif %}
-{% endblock content %}
