--- packages/fkweb/templates/fkvod/video_list.html
@@ -1,181 +1,0 @@
-{% extends "base_generic.html" %}
-{% load fk_utils %}
-
-{% block style %}
-<style>
-
-.video_container {
-    clear: left;
-}
-
-.videolist_container {
-}
-
-.video_container ul {
-    list-style: none;
-    margin: 5px;
-    padding: 5px;
-    /*margin: 15px 15px 0 0;*/
-    border-radius: 8px;
-    background-color:#eaeaea;
-    /*border: 3px solid;
-    border-color:#d8d8d8;*/
-}
-
-.video_container img {
-    float: left;
-    padding: 0px 8px 4px 4px;
-}
-.video_container img {
-    height: auto;
-    width: 300px;
-}
-
-.video_container .name {
-    font-weight: bold;
-}
-
-.video_container .description {
-    font-size: 75%;
-}
-
-.video_container .duration {
-    font-size: 67%;
-}
-
-.video_container .editor {
-    font-size: 75%;
-}
-
-.video_container li {
-    padding:0;
-    margin:0;
-}
-
-</style>
-
-{% endblock style %}
-
-{% block content %}
-{% if org %}
-  {% with name=org.name homepage=org.homepage %}
-  {% if org.description %}
-  <p>{{ org.description }}</p>
-  {% endif %}
-  {% if org.homepage %}
-  <p>{% blocktrans %}Visit <a href="{{ homepage }}">web pages for {{ name }}</a>{% endblocktrans %}</p>
-  {% endif %}
-
-  {% endwith %}
-{% endif %}
-
-{% if page.paginator.num_pages > 1 %}
-<div class="navigation_container">
-    <div class="navigation">
-        <div class="inner">
-        {% if page.has_previous %}
-        <a href="?page={{ page.previous_page_number }}{{ url_query_postfix }}"> {% trans "Previous" %}</a>
-        {% else %}
-        {% trans "First" %}
-        {% endif %}
-        </div>
-    </div>
-
-    <div class="navigation">
-        <div class="inner">
-          {% with number=page.number num_pages=page.paginator.num_pages %}
-            {% blocktrans %}{{ number }} of {{ num_pages }}{% endblocktrans %}
-	  {% endwith  %}
-        </div>
-    </div>
-
-    <div class="navigation">
-        <div class="inner">
-        {% if page.has_next %}
-        <a href="?page={{ page.next_page_number }}{{ url_query_postfix }}">{% trans "Next" %}</a>
-        {% else %}
-        {% trans "Last" %}
-        {% endif %}
-        </div>
-    </div>
-</div>
-{% endif %}
-
-  <div class="search">
-      <form>
-          <input name="q" value="{{ search_query }}" placeholder="Search videos" type="search">
-          <input type="submit" value="Search">
-      </form>
-  </div>
-
-<div class="videolist_container">
-{% for video in videos %}
-<div class="video_container">
-    <ul>
-        <li>
-            <a href="{% url 'vod-video-detail' video.id %}">
-                <img src="{{ video.large_thumbnail_url }}" class="small_thumbnail"/>
-            </a>
-        </li>
-        <li class="header">
-            <a href="{% url 'vod-video-detail' video.id %}"><span class="name">{{ video.name }}</span></a>
-            <span class="duration">({{ video.duration|truncate_duration }})</span>
-        </li>
-        <li class="editor">
-        {% if video.organization %}
-            <a href="{% url 'vod-org-video-list' video.organization.id %}">
-            <i>{{ video.organization }}</i>
-            </a>
-	{% else %}
-            <i>n/a</i>
-	{% endif %}
-        </li>
-        <li class="description">
-        {% if video.header %}
-        <i>
-            {{ video.header }}
-        </i></li>
-        {% endif %}
-
-    </ul>
-</div>
-{% endfor %}
-</div>
-
-{% if page.paginator.num_pages > 1 %}
-<div class="navigation_container">
-    <div class="navigation">
-        <div class="inner">
-        {% if page.has_previous %}
-        <a href="?page={{ page.previous_page_number }}{{ url_query_postfix }}"> {% trans "Previous" %}</a>
-        {% else %}
-        {% trans "First" %}
-        {% endif %}
-        </div>
-    </div>
-
-    <div class="navigation">
-        <div class="inner">
-          {% with number=page.number num_pages=page.paginator.num_pages %}
-            {% blocktrans %}{{ number }} of {{ num_pages }}{% endblocktrans %}
-	  {% endwith  %}
-        </div>
-    </div>
-
-    <div class="navigation">
-        <div class="inner">
-        {% if page.has_next %}
-        <a href="?page={{ page.next_page_number }}{{ url_query_postfix }}">{% trans "Next" %}</a>
-        {% else %}
-        {% trans "Last" %}
-        {% endif %}
-        </div>
-    </div>
-</div>
-{% endif %}
-
-<div class="navigation_container">
-    <div class="navigation">[<a href="./rss?{{ request.GET.urlencode }}">RSS</a>]</div>
-</div>
-
-{% endblock content %}
