--- packages/fkweb/templates/fkvod/video_list.rss
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<rss version="2.0"
-  xmlns:itunes="http://www.itunes.com/dtds/podcast-1.0.dtd"
-  xmlns:atom="http://www.w3.org/2005/Atom"
->
-
-  <channel>
-    <title>Innslag på Frikanalen</title>
-    <link>https://frikanalen.no/</link>
-    <atom:link href="https://frikanalen.no/video/rss" rel="self" type="application/rss+xml" />
-    <description>Video recordings published via the Norwegian Open Channel Frikanalen.</description>
-
-    <copyright>(c) 2009-2018 Frikanalen and contributors</copyright>
-    <language>nb</language>
-    <itunes:author>Frikanalen</itunes:author>
-    <itunes:subtitle>Collected recordings</itunes:subtitle>
-    <itunes:summary>For Free.</itunes:summary>
-    <itunes:owner>
-        <itunes:name>Frikanalen</itunes:name>
-        <itunes:email>post@frikanalen.no</itunes:email>
-    </itunes:owner>
-    <category>Cartoons</category>
-    <itunes:category text="Technology">
-        <itunes:category text="Software How-To" />
-    </itunes:category>
-
-    <itunes:keywords>Frikanalen, Norway</itunes:keywords>
-    <image>
-        <title>Frikanalen</title>
-        <url>http://www.frikanalen.no/images/logo_small.png</url>
-        <link>https://frikanalen.no/video/rss</link>
-    </image>
-    <itunes:image href="http://www.frikanalen.no/images/logo_small.png"/>
-
-{% for video in videos %}
-    <item>
-      <title>{{ video.name }}</title>
-       <pubDate>{{ video.uploaded_time }}</pubDate>
-      <itunes:duration>{{ video.duration }}</itunes:duration>
-      <enclosure url="{{ video.ogv_url }}"/>
-    </item>
-
-{% endfor %}
-
-  </channel>
-</rss>
