--- packages/fkweb/templates/fkvod/vod_widget.html
@@ -1,32 +1,0 @@
-<!-- VOD Player -->
-<div class="player-container">
-   {% if video_error %}
-   <div id="player-error-container" style="background-image: url('{{ video.large_thumbnail_url }}');">
-       <div id="player-error-message"><h1>{{ video_error }}</h1>{{ video_error_explanation }}</div>
-   </div>
-   {% elif video.vod_files %}
-   <video class="player" poster="{{ video.large_thumbnail_url }}" controls>
-       {% for vod_file in video.vod_files %}
-         <source src="{{ vod_file.url }}" type="{{ vod_file.mime_type }}"/>
-       {% endfor %}
-       {% if video.ogv_url %}
-         <applet code="com.fluendo.player.Cortado.class" archive="http://theora.org/cortado.jar">
-           <param name="url" value="{{ video.ogv_url }}">
-       {% endif %}
-       <div id="player-error-container" style="background-image: url('{{ video.large_thumbnail_url }}');">
-           <div id="player-error-message">
-               <h1>Browser does not support HTML5 video</h1>
-               Please use Chrome, Opera, Firefox or any other browser which supports Ogg Theora.
-           </div>
-       </div>
-       {% if video.ogv_url %}
-         </applet>
-       {% endif %}
-   </video>
-   {% else %}
-   <div id="player-error-container" style="background-image: url('{{ video.large_thumbnail_url }}');">
-       <div id="player-error-message"><h1>Video not available</h1></div>
-   </div>
-   {% endif %}
-</div>
-<!-- VOD Player end -->
